Generalised image tiles to N dimensions
Added the start of theory for generating rotated sets of tiles

Expressed constraints in terms of a struct, allowing types to be erased,
avoiding having to write tile types all over the shop.

Changed domain definition to use a type alias for entries rather than a
newtype struct and renamed a few bits.