Docs + Removed Heuristics
## Documentation
Went on a bit of a documentation binge, added top level docs for:
- `Domain`s
- `AC3Error`s
- `Space`s
- `Cartesian` and `Wrapped` grids
- `First`, `Uniform` and `Weighted` sampling
- 'Backtrack', `Naive` and `Restart` searches

## Abandoned Heuristics
Stopped working on and removed customisable heuristics.
This was because the current heuristics used get a good balance between
being effective and being simple to understand and it's unlikely that
anyone will want to change the heuristic without changing the library itself.